feat: filter makers by user.id

+++ src/app/api/markers/[id]/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse } from 'next/server'
 import dbConnect from '@@/lib/mongodb'
 import Marker from '@@/models/MarkerModel'
+import { auth } from '@@/lib/auth'
 
 export async function DELETE(
 	req: Request,
@@ -9,10 +10,24 @@ export async function DELETE(
 	const { id } = params
 	try {
 		await dbConnect()
-		const marker = await Marker.findById(id)
+		const session = await auth()
+		if (!session) {
+			return NextResponse.json(
+				{ message: 'Необходима авторизация' },
+				{ status: 401 }
+			)
+		}
 
+		const marker = await Marker.findById(id)
 		if (!marker) {
 			return NextResponse.json({ message: 'Метка не найдена' }, { status: 404 })
+		}
+
+		if (marker.user.toString() !== session.user.id) {
+			return NextResponse.json(
+				{ message: 'У вас нет прав для удаления этой метки' },
+				{ status: 403 }
+			)
 		}
 
 		await marker.deleteOne()
@@ -38,16 +53,29 @@ export async function PATCH(
 
 	try {
 		await dbConnect()
-		const updatedMarker = await Marker.findByIdAndUpdate(id, data, {
-			new: true
-		})
+		const session = await auth()
+		if (!session) {
+			return NextResponse.json(
+				{ message: 'Необходима авторизация' },
+				{ status: 401 }
+			)
+		}
 
+		const updatedMarker = await Marker.findById(id)
 		if (!updatedMarker) {
 			return NextResponse.json({ message: 'Метка не найдена' }, { status: 404 })
 		}
 
+		if (updatedMarker.user.toString() !== session.user.id) {
+			return NextResponse.json(
+				{ message: 'У вас нет прав для обновления этой метки' },
+				{ status: 403 }
+			)
+		}
+		const marker = await Marker.findByIdAndUpdate(id, data, { new: true })
+
 		return NextResponse.json(
-			{ message: 'Метка успешно обновлена', data: updatedMarker },
+			{ message: 'Метка успешно обновлена', data: marker },
 			{ status: 200 }
 		)
 	} catch (error) {
